This didn’t work
Instead I made a basic module to figure out how it worked. Then realized the button form was reloading before it was logging in. Fixed it for extended login.

There’s a couple issues still:
If you login to an account that has Twitter and Google, you won’t have access to Google API this way.

Eventually I can create a Firebase script that gets auth on the server and keeps authentication on the server.

At the moment, I don’t need server access to Google APIs so I’d rather not (for security issues)